// apps/api/src/modules/auth/domain/audit-event.enum.ts
export enum AuditEvent {
  USER_REGISTERED = 'USER_REGISTERED',
  USER_CREATED_BY_ADMIN = 'USER_CREATED_BY_ADMIN',
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILURE = 'LOGIN_FAILURE',
  LOGOUT = 'LOGOUT',
  TOKEN_REFRESH = 'TOKEN_REFRESH',
  ACCOUNT_LOCKED = 'ACCOUNT_LOCKED',
  REFRESH_TOKEN_REUSE_DETECTED = 'REFRESH_TOKEN_REUSE_DETECTED',
  PATIENT_CREATED = 'PATIENT_CREATED',
  PATIENT_LISTED = 'PATIENT_LISTED',
  PATIENT_SEARCHED = 'PATIENT_SEARCHED',
  PATIENT_VIEWED = 'PATIENT_VIEWED',
  PATIENT_UPDATED = 'PATIENT_UPDATED',
  PATIENT_DEACTIVATED = 'PATIENT_DEACTIVATED',
  PATIENT_EXPORTED = 'PATIENT_EXPORTED',
  MEDICAL_RECORD_CREATED = 'MEDICAL_RECORD_CREATED',
  MEDICAL_RECORD_VIEWED = 'MEDICAL_RECORD_VIEWED',
  MEDICAL_RECORD_LISTED = 'MEDICAL_RECORD_LISTED',
  MEDICAL_RECORD_UPDATED = 'MEDICAL_RECORD_UPDATED',
  MEDICAL_RECORD_SIGNED = 'MEDICAL_RECORD_SIGNED',
  MEDICAL_RECORD_ADDENDUM_CREATED = 'MEDICAL_RECORD_ADDENDUM_CREATED',
  CID10_SEARCHED = 'CID10_SEARCHED',
  NOTIFICATION_CREATED = 'NOTIFICATION_CREATED',
  NOTIFICATION_QUEUED = 'NOTIFICATION_QUEUED',
  NOTIFICATION_SENT = 'NOTIFICATION_SENT',
  NOTIFICATION_FAILED = 'NOTIFICATION_FAILED',
  NOTIFICATION_DASHBOARD_VIEWED = 'NOTIFICATION_DASHBOARD_VIEWED',
  NOTIFICATION_OPT_OUT_UPDATED = 'NOTIFICATION_OPT_OUT_UPDATED',
  DOCUMENT_UPLOAD_URL_CREATED = 'DOCUMENT_UPLOAD_URL_CREATED',
  DOCUMENT_UPLOAD_CONFIRMED = 'DOCUMENT_UPLOAD_CONFIRMED',
  DOCUMENT_UPLOAD_REJECTED = 'DOCUMENT_UPLOAD_REJECTED',
  DOCUMENT_VIEW_URL_CREATED = 'DOCUMENT_VIEW_URL_CREATED',
  DOCUMENT_LISTED = 'DOCUMENT_LISTED',
  DOCUMENT_SOFT_DELETED = 'DOCUMENT_SOFT_DELETED',
  CLINIC_CREATED = 'CLINIC_CREATED',
  TENANT_CONTEXT_RESOLVED = 'TENANT_CONTEXT_RESOLVED',
  APPOINTMENT_CREATED = 'APPOINTMENT_CREATED',
  APPOINTMENT_LISTED = 'APPOINTMENT_LISTED',
  APPOINTMENT_VIEWED = 'APPOINTMENT_VIEWED',
  APPOINTMENT_UPDATED = 'APPOINTMENT_UPDATED',
  APPOINTMENT_CANCELLED = 'APPOINTMENT_CANCELLED',
  APPOINTMENT_CONCLUDED = 'APPOINTMENT_CONCLUDED',
  SCHEDULE_BLOCKED = 'SCHEDULE_BLOCKED',
  SCHEDULE_UNBLOCKED = 'SCHEDULE_UNBLOCKED',
  FINANCIAL_ENTRY_CREATED = 'FINANCIAL_ENTRY_CREATED',
  FINANCIAL_ENTRY_LISTED = 'FINANCIAL_ENTRY_LISTED',
  FINANCIAL_ENTRY_RECEIVED = 'FINANCIAL_ENTRY_RECEIVED',
  FINANCIAL_ENTRY_CANCELLED = 'FINANCIAL_ENTRY_CANCELLED',
  FINANCIAL_DASHBOARD_VIEWED = 'FINANCIAL_DASHBOARD_VIEWED',
  FINANCIAL_REPORT_VIEWED = 'FINANCIAL_REPORT_VIEWED',
  FINANCIAL_REPORT_EXPORTED = 'FINANCIAL_REPORT_EXPORTED',
  FINANCIAL_SERVICE_SAVED = 'FINANCIAL_SERVICE_SAVED',
  FINANCIAL_INSTITUTION_SAVED = 'FINANCIAL_INSTITUTION_SAVED',
  FINANCIAL_RECURRENCE_SAVED = 'FINANCIAL_RECURRENCE_SAVED',
  FINANCIAL_RECURRENCE_MATERIALIZED = 'FINANCIAL_RECURRENCE_MATERIALIZED',
  TELEMEDICINE_ROOM_CREATED = 'TELEMEDICINE_ROOM_CREATED',
  TELEMEDICINE_ROOM_JOINED = 'TELEMEDICINE_ROOM_JOINED',
  TELEMEDICINE_ROOM_ENDED = 'TELEMEDICINE_ROOM_ENDED',
  WOUND_CREATED = 'WOUND_CREATED',
  WOUND_UPDATED = 'WOUND_UPDATED',
  WOUND_ASSESSMENT_CREATED = 'WOUND_ASSESSMENT_CREATED',
}